requirements.txt
gdown==5.1.0
imageio==2.21.2
matplotlib==3.7.5
numpy==1.24.4
Pillow==10.4.0
scipy==1.10.1
torch
torchvision
tqdm==4.66.1
transformers==4.21.3
timm==0.6.7
scikit-learn==1.3.2
pandas==2.0.3
einops==0.6.1
PyYAML==6.0.1

